+++ test/e2e/autoscaling/custom_metrics_stackdriver_autoscaling.go
@@ -1,3 +1,5 @@
+// +build !providerless
+
 /*
 Copyright 2017 The Kubernetes Authors.
