+++ test/e2e/framework/providers/gce/firewall_test.go
@@ -1,3 +1,5 @@
+// +build !providerless
+
 /*
 Copyright 2018 The Kubernetes Authors.
