+++ test/e2e/instrumentation/monitoring/custom_metrics_deployments.go
@@ -1,3 +1,5 @@
+// +build !providerless
+
 /*
 Copyright 2017 The Kubernetes Authors.
