+++ test/e2e/scheduling/nvidia-gpus.go
@@ -1,3 +1,5 @@
+// +build !providerless
+
 /*
 Copyright 2017 The Kubernetes Authors.
