+++ test/e2e/scheduling/ubernetes_lite_volumes.go
@@ -1,3 +1,5 @@
+// +build !providerless
+
 /*
 Copyright 2017 The Kubernetes Authors.
